Clarification
