--- Sudoku.PierreLouis/dl/dlxsudoku/__init__.py
@@ -1,1 +1,0 @@
-from .sudoku import Sudoku
