--- Sudoku.PierreLouis/dl/tests/hard.sud
@@ -1,10 +1,0 @@
-# Hard 1
-****5**7*
-*5**1*4*2
-871******
-*6*42***5
-**4***6**
-2***76*1*
-******851
-5*7*4**3*
-*1**9****
