--- Sudoku.PierreLouis/dl/tests/medium.sud
@@ -1,10 +1,0 @@
-# 157 - Medelklurigt
-*72****6*
-***72*9*4
-*9*1****2
-*******4*
-82*4*71**
-**9*6*8**
-***9**6**
-**3*72*9*
-*6*843*7*
