--- Sudoku.PierreLouis/dl/tests/medium_sol.sud
@@ -1,10 +1,0 @@
-# 157 - Medelklurigt
-172394568
-358726914
-694158732
-715289346
-826437159
-439561827
-247915683
-583672491
-961843275
