--- Sudoku.PierreLouis/dl/tests/simple_sol.sud
@@ -1,9 +1,0 @@
-138467259
-924518376
-567392148
-351946827
-472851693
-896273415
-785624931
-213789564
-649135782
