--- Sudoku.PierreLouis/dl/tests/very_hard.sud
@@ -1,10 +1,0 @@
-# Very hard 278
-***4*****
-**9******
-*3**7****
-***7****8
-****5*32*
-4**86***5
-5*3****8*
-7983**4**
-**6**9***
