--- Sudoku.PierreLouis/dl/tests/very_hard_sol.sud
@@ -1,10 +1,0 @@
-# Very hard 278
-652431897
-179628534
-834975261
-365792148
-987154326
-421863975
-513247689
-798316452
-246589713
