+++ Sudoku.PierreLouis/dlxsudoku/__init__.py
@@ -1,0 +1,1 @@
+from .sudoku import Sudoku
