+++ Sudoku.PierreLouis/sudokus/grids_2.sud
@@ -1,0 +1,4 @@
+Grid 01
+600120384008459072000006005000264030070080006940003000310000050089700000502000190
+Grid 50
+300200000000107000706030500070009080900020004010800050009040301000702000000008006
