+++ Sudoku.PierreLouis/tests/hard_sol.sud
@@ -1,0 +1,10 @@
+# Hard 1
+642359178
+953817462
+871264593
+169423785
+734581629
+285976314
+496732851
+527148936
+318695247
