+++ Sudoku.PierreLouis/tests/medium_sol.sud
@@ -1,0 +1,10 @@
+# 157 - Medelklurigt
+172394568
+358726914
+694158732
+715289346
+826437159
+439561827
+247915683
+583672491
+961843275
