+++ Sudoku.PierreLouis/tests/simple.sud
@@ -1,0 +1,9 @@
+*3*467*5*
+92**1***6
+*673**148
+3*1**6*27
+4**85*6**
+*9*2**4**
+**5624**1
+2*3***5*4
+*4**3*7*2
