+++ Sudoku.PierreLouis/tests/simple_flat.sud
@@ -1,0 +1,1 @@
+030467050920010006067300148301006027400850600090200400005624001203000504040030702
